+++ pkg/apis/minio.min.io/v1/helper.go
@@ -139,11 +139,6 @@ func (t *Tenant) ConsoleExternalCert() bool {
 
 // AutoCert is enabled by default, otherwise we return the user provided value
 func (t *Tenant) AutoCert() bool {
-	// AutoCertEnabled will take priority over RequestAutoCert and
-	// will be removed in the future
-	if t.Status.Certificates.AutoCertEnabled != nil {
-		return *t.Status.Certificates.AutoCertEnabled
-	}
 	if t.Spec.RequestAutoCert == nil {
 		return true
 	}
