+++ pkg/resources/statefulsets/minio-statefulset.go
@@ -195,7 +195,7 @@ func volumeMounts(t *miniov2.Tenant, pool *miniov2.Pool) (mounts []corev1.Volume
 		}
 	}
 
-	if t.AutoCert() || t.ExternalCert() {
+	if t.AutoCert() || t.ExternalCert() || t.HasKESEnabled() {
 		mounts = append(mounts, corev1.VolumeMount{
 			Name:      t.MinIOTLSSecretName(),
 			MountPath: miniov2.MinIOCertPath,
@@ -286,35 +286,6 @@ func NewForMinIOPool(t *miniov2.Tenant, wsSecret *v1.Secret, pool *miniov2.Pool,
 				Items: serverCertPaths,
 			},
 		})
-	}
-
-	// External Client certificates will have priority over AutoCert generated certificates
-	if t.ExternalClientCert() {
-		clientCertSecret = t.Spec.ExternalClientCertSecret.Name
-		// This covers both secrets of type "kubernetes.io/tls" and
-		// "cert-manager.io/v1alpha2" because of same keys in both.
-		if t.Spec.ExternalClientCertSecret.Type == "kubernetes.io/tls" || t.Spec.ExternalClientCertSecret.Type == "cert-manager.io/v1alpha2" {
-			clientCertPaths = []corev1.KeyToPath{
-				{Key: "tls.crt", Path: "client.crt"},
-				{Key: "tls.key", Path: "client.key"},
-			}
-		}
-	} else if t.AutoCert() {
-		clientCertSecret = t.MinIOClientTLSSecretName()
-	}
-
-	// KES External certificates will have priority over AutoCert generated certificates
-	if t.KESExternalCert() {
-		kesCertSecret = t.Spec.KES.ExternalCertSecret.Name
-		// This covers both secrets of type "kubernetes.io/tls" and
-		// "cert-manager.io/v1alpha2" because of same keys in both.
-		if t.Spec.KES.ExternalCertSecret.Type == "kubernetes.io/tls" || t.Spec.KES.ExternalCertSecret.Type == "cert-manager.io/v1alpha2" {
-			KESCertPath = []corev1.KeyToPath{
-				{Key: "tls.crt", Path: "CAs/kes.crt"},
-			}
-		}
-	} else if t.AutoCert() {
-		kesCertSecret = t.KESTLSSecretName()
 	}
 
 	if t.ExternalCert() {
@@ -456,8 +427,37 @@ func NewForMinIOPool(t *miniov2.Tenant, wsSecret *v1.Secret, pool *miniov2.Pool,
 		},
 	}...)
 
-	// Add SSL volume from SSL secret to the podVolumes
-	if t.TLS() && t.HasKESEnabled() {
+	// If KES is enable mount TLS certificate secrets
+	if t.HasKESEnabled() {
+		// External Client certificates will have priority over AutoCert generated certificates
+		if t.ExternalClientCert() {
+			clientCertSecret = t.Spec.ExternalClientCertSecret.Name
+			// This covers both secrets of type "kubernetes.io/tls" and
+			// "cert-manager.io/v1alpha2" because of same keys in both.
+			if t.Spec.ExternalClientCertSecret.Type == "kubernetes.io/tls" || t.Spec.ExternalClientCertSecret.Type == "cert-manager.io/v1alpha2" {
+				clientCertPaths = []corev1.KeyToPath{
+					{Key: "tls.crt", Path: "client.crt"},
+					{Key: "tls.key", Path: "client.key"},
+				}
+			}
+		} else {
+			clientCertSecret = t.MinIOClientTLSSecretName()
+		}
+
+		// KES External certificates will have priority over AutoCert generated certificates
+		if t.KESExternalCert() {
+			kesCertSecret = t.Spec.KES.ExternalCertSecret.Name
+			// This covers both secrets of type "kubernetes.io/tls" and
+			// "cert-manager.io/v1alpha2" because of same keys in both.
+			if t.Spec.KES.ExternalCertSecret.Type == "kubernetes.io/tls" || t.Spec.KES.ExternalCertSecret.Type == "cert-manager.io/v1alpha2" {
+				KESCertPath = []corev1.KeyToPath{
+					{Key: "tls.crt", Path: "CAs/kes.crt"},
+				}
+			}
+		} else {
+			kesCertSecret = t.KESTLSSecretName()
+		}
+
 		podVolumeSources = append(podVolumeSources, []corev1.VolumeProjection{
 			{
 				Secret: &corev1.SecretProjection{
